backend/requirements.txt
requests 
packaging
fastapi
uvicorn 
PyGithub
python-dotenv
pydantic

